feat: move home to own subdirectory.
logion-network/logion-internal#1197
